Hacked beam_map to get evaluate_detections working
